Restructure from single file
